Move coordinate_is_on_solar_disk to coordinates utils
