+++ sunpy/map/tests/test_maputils.py
@@ -10,7 +10,8 @@ from astropy.tests.helper import assert_quantity_allclose
 import sunpy.map
 from sunpy.coordinates import HeliographicStonyhurst
 from sunpy.coordinates.frames import HeliographicCarrington
-from sunpy.coordinates.utils import GreatArc
+from sunpy.coordinates.utils import GreatArc,coordinate_is_on_solar_disk,solar_angular_radius
+from sunpy.util.exceptions import SunpyDeprecationWarning
 from sunpy.map.maputils import (
     _verify_coordinate_helioprojective,
     all_coordinates_from_map,
@@ -20,14 +21,13 @@ from sunpy.map.maputils import (
     contains_full_disk,
     contains_limb,
     contains_solar_center,
-    coordinate_is_on_solar_disk,
     is_all_off_disk,
     is_all_on_disk,
     map_edges,
     on_disk_bounding_coordinates,
     pixelate_coord_path,
     sample_at_coords,
-    solar_angular_radius,
+
 )
 
 
@@ -250,7 +250,7 @@ def test_verify_coordinate_helioprojective(aia171_test_map, all_off_disk_map, al
 
 def test_functions_raise_non_frame_coords(non_helioprojective_skycoord):
     with pytest.raises(ValueError, match=r"ICRS, .* Helioprojective"):
-        solar_angular_radius(non_helioprojective_skycoord)
+        coordinate_is_on_solar_disk(non_helioprojective_skycoord)
     with pytest.raises(ValueError, match=r"ICRS, .* Helioprojective"):
         coordinate_is_on_solar_disk(non_helioprojective_skycoord)
